Revert "Edc gem"
