--- Gemfile
@@ -1,4 +1,0 @@
-source 'https://rubygems.org'
-
-# Specify your gem's dependencies in rails_memory_bloat.gemspec
-gemspec
