--- bin/console
@@ -1,14 +1,0 @@
-#!/usr/bin/env ruby
-
-require "bundler/setup"
-require "rails_memory_bloat"
-
-# You can add fixtures and/or initialization code here to make experimenting
-# with your gem easier. You can also use a different console, if you like.
-
-# (If you use this, don't forget to add pry to your Gemfile!)
-# require "pry"
-# Pry.start
-
-require "irb"
-IRB.start
