--- lib/rails_memory_bloat/version.rb
@@ -1,3 +1,0 @@
-module RailsMemoryBloat
-  VERSION = "0.1.0"
-end
